+++ src/hub/uninstall-task.ts
@@ -7,7 +7,7 @@ import { tkn } from '../tkn';
 import { HubTaskUninstall } from './hub-common';
 import * as vscode from 'vscode';
 import { getStderrString } from '../util/stderrstring';
-import { Command } from '../util/command';
+import { Command } from '../cli-command';
 
 const uninstallTaskEmitter = new vscode.EventEmitter<HubTaskUninstall>();
 export const uninstallTaskEvent = uninstallTaskEmitter.event;
