+++ src/tekton/clustertask.ts
@@ -5,8 +5,8 @@
 
 import { TektonItem } from './tektonitem';
 import { CliCommand } from '../cli';
-import { Command } from '../util/command';
 import { TektonNode } from '../tree-view/tekton-node';
+import { Command } from '../cli-command';
 
 export class ClusterTask extends TektonItem {
