+++ src/tekton/task.ts
@@ -7,8 +7,9 @@ import { TektonItem } from './tektonitem';
 import { window } from 'vscode';
 import * as cliInstance from '../cli';
 import { startTask } from './starttask';
-import { Command } from '../util/command';
 import { TektonNode } from '../tree-view/tekton-node';
+import { Command } from '../cli-command';
+
 
 export class Task extends TektonItem {
