+++ src/tekton/tektonitem.ts
@@ -9,9 +9,10 @@ import { workspace, window } from 'vscode';
 import { tektonFSUri } from '../util/tekton-vfs';
 import { TknResourceItem } from './webviewstartpipeline';
 import { telemetryLogCommand, telemetryLogError } from '../telemetry';
-import { Command } from '../util/command';
 import { ContextType } from '../context-type';
 import { TektonNode } from '../tree-view/tekton-node';
+import { Command } from '../cli-command';
+
 
 export const errorMessage = {
   Pipeline: 'You need at least one Pipeline available. Please create new Tekton Pipeline and try again.',
