+++ src/util/tknversion.ts
@@ -3,8 +3,9 @@
  *  Licensed under the MIT License. See LICENSE file in the project root for license information.
  *-----------------------------------------------------------------------------------------------*/
 
+import { Command } from '../cli-command';
 import { tkn } from '../tkn';
-import { Command } from './command';
+
 
 export interface TknVersion {
   trigger?: string;
